Move platform/auth into infra/auth
Change-type: patch
