+++ src/routes/access.ts
@@ -3,7 +3,7 @@ import type { Request, Response } from 'express';
 import { sbvrUtils, errors } from '@balena/pinejs';
 import * as semver from 'balena-semver';
 
-import { reqHasPermission } from '../platform/auth';
+import { reqHasPermission } from '../infra/auth/auth';
 import { captureException } from '../infra/error-handling';
 
 const { UnauthorizedError } = errors;
